add support to generate xmake project on multiplatforms

+++ src/main/kotlin/io/xmake/project/XMakeModuleBuilder.kt
@@ -1,5 +1,8 @@
 package io.xmake.project
 
+import com.intellij.execution.configurations.GeneralCommandLine
+import com.intellij.execution.process.ProcessNotCreatedException
+import com.intellij.execution.wsl.WSLDistribution
 import com.intellij.ide.util.projectWizard.*
 import com.intellij.openapi.Disposable
 import com.intellij.openapi.diagnostic.Logger
@@ -9,14 +12,22 @@ import com.intellij.openapi.roots.ModifiableRootModel
 import com.intellij.openapi.util.Disposer
 import com.intellij.openapi.util.io.FileUtil
 import com.intellij.openapi.vfs.LocalFileSystem
+import com.intellij.ssh.config.unified.SshConfig
+import io.xmake.project.toolkit.ToolkitHostType.*
+import io.xmake.run.XMakeRunConfiguration
+import io.xmake.run.XMakeRunConfigurationType
 import io.xmake.utils.SystemUtils
-import io.xmake.utils.ioRunvInPool
+import io.xmake.utils.execute.*
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.runBlocking
 import java.io.File
 
 
 class XMakeModuleBuilder : ModuleBuilder() {
-    var configurationData: XMakeConfigData? = null
+    lateinit var configurationData: XMakeConfigData
 
+    private val scope = CoroutineScope(Dispatchers.Default)
 
     override fun isSuitableSdkType(sdkType: SdkTypeId?): Boolean = true
 
@@ -38,22 +49,57 @@ class XMakeModuleBuilder : ModuleBuilder() {
          * @note we muse use ioRunv instead of Runv to read all output, otherwise it will wait forever on windows
          */
         val tmpdir = "$contentEntryPath.dir"
-        ioRunvInPool(
-            listOf(
-                SystemUtils.xmakeProgram,
-                "create",
-                "-P",
-                tmpdir,
-                "-l",
-                configurationData?.languagesModel.toString(),
-                "-t",
-                configurationData?.kindsModel.toString()
-            )
+
+        val dir = when(configurationData.toolkit!!.host.type) {
+            LOCAL -> tmpdir
+            WSL -> configurationData.remotePath
+            SSH -> configurationData.remotePath
+        }
+
+        Log.debug("dir: $dir")
+
+        val command = listOf(
+            SystemUtils.xmakeProgram,
+            "create",
+            "-P",
+            dir,
+            "-l",
+            configurationData.languagesModel,
+            "-t",
+            configurationData.kindsModel
         )
-        val tmpFile = File(tmpdir)
-        if (tmpFile.exists()) {
-            tmpFile.copyRecursively(File(contentEntryPath), true)
-            tmpFile.deleteRecursively()
+
+        val commandLine: GeneralCommandLine = GeneralCommandLine(command)
+//            .withWorkDirectory(workDir)
+            .withCharset(Charsets.UTF_8)
+            .withParentEnvironmentType(GeneralCommandLine.ParentEnvironmentType.CONSOLE)
+        val results = try {
+            val (result, exitCode) = runBlocking(Dispatchers.IO) {
+                return@runBlocking runProcess(commandLine.createProcess(configurationData.toolkit!!))
+            }
+            result.getOrDefault("").split(Regex("\\s+"))
+        } catch (e: ProcessNotCreatedException) {
+            emptyList()
+        }
+
+        Log.info("results: $results")
+
+        with(configurationData.toolkit!!) {
+            when(host.type) {
+                LOCAL -> {
+                    val tmpFile = File(tmpdir)
+                    if (tmpFile.exists()) {
+                        tmpFile.copyRecursively(File(contentEntryPath), true)
+                        tmpFile.deleteRecursively()
+                    }
+                }
+                WSL -> {
+                    syncProjectByWslSync(scope, rootModel.project, host.target as WSLDistribution, configurationData.remotePath!!, SyncDirection.UPSTREAM_TO_LOCAL)
+                }
+                SSH -> {
+                    syncProjectBySftp(scope, rootModel.project, host.target as SshConfig, configurationData.remotePath!!, SyncDirection.UPSTREAM_TO_LOCAL)
+                }
+            }
         }
     }
